refactor(security): share credential filter boundary
