+++ nemoclaw/tsconfig.shared.json
@@ -6,6 +6,7 @@
   },
   "include": [
     "src/shared/banner-boundary.cts",
+    "src/shared/credential-filter-boundary.cts",
     "src/shared/openshell-policy-boundary.cts",
     "src/shared/sandbox-name.cts",
     "src/shared/snapshot-sanitizer-boundary.cts"
