+++ vitest.config.ts
@@ -21,6 +21,9 @@ const { isCi, silent } = resolveVitestFeedback();
 const LIVE_E2E_PROJECT_TIMEOUT_MS = 30 * 60 * 1000;
 const runLiveE2E = shouldRunLiveE2E();
 const canonicalBannerBoundary = path.resolve("nemoclaw/src/shared/banner-boundary.cts");
+const canonicalCredentialFilterBoundary = path.resolve(
+  "nemoclaw/src/shared/credential-filter-boundary.cts",
+);
 const canonicalOpenShellPolicyBoundary = path.resolve(
   "nemoclaw/src/shared/openshell-policy-boundary.cts",
 );
@@ -35,6 +38,10 @@ const canonicalSourceAliases = [
   {
     find: /^.*banner-boundary\.cjs$/,
     replacement: canonicalBannerBoundary,
+  },
+  {
+    find: /^.*credential-filter-boundary\.cjs$/,
+    replacement: canonicalCredentialFilterBoundary,
   },
   {
     find: /^.*openshell-policy-boundary\.cjs$/,
